requirements.txt
cohere
langchain
openai
pandas
gitpython
lark
cython
tiktoken
fastapi
python-multipart
uvicorn
pyowm
python-dotenv
termcolor
gunicorn
schedule
qdrant-client
apscheduler
nltk
cachetools
jsonpatch
pymongo
pytest
pytest-asyncio
motor